- Merged SharedRender into SharedGame and deleted SharedRender - Moved around some includes and forward declared asome things to open the possibility of including SharedGame into the networking framework - Full loopback support for the networking menu(s) and option(s)

+++ Flap/Main/Host.cpp
@@ -70,7 +70,7 @@ void Host::GenAssAndSendSpecMess(SharedNetwork::SpecialMessage _specialMessage, 
 	if (_addressOrPort != ULONG_MAX)
 	{
 #ifdef SAME_SYSTEM_NETWORK
-		m_commSockAddrIn.sin_port = _addressOrPort;
+		m_commSockAddrIn.sin_port = static_cast<unsigned short>(_addressOrPort);
 #else !SAME_SYSTEM_NETWORK
 		m_commSockAddrIn.sin_addr.S_un.S_addr = _addressOrPort;
 #endif SAME_SYSTEM_NETWORK
